Refactor: Adjust macro import paths for clarity and modularity; update all references accordingly.
