+++ src/f64/math.rs
@@ -1,5 +1,5 @@
 use super::{GuardedF64, UnguardedF64};
-use crate::macros::math;
+use crate::macros::math::math;
 
 math!(
     (GuardedF64, UnguardedF64)
