+++ src/f64/ops_binary.rs
@@ -1,5 +1,5 @@
 use super::{GuardedF64, UnguardedF64};
-use crate::macros::binary_operation;
+use crate::macros::ops_binary::binary_operation;
 use std::ops::{Add, Div, Mul, Rem, Sub};
 
 binary_operation!(
