+++ src/lib.rs
@@ -3,9 +3,7 @@
 mod error;
 mod f32;
 mod f64;
-
-#[macro_use]
-mod macros;
+pub(crate) mod macros;
 
 pub use error::Error as FloatError;
 pub use f32::{GuardedF32, UnguardedF32};
